optimize alert delete

+++ app/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState, useCallback } from "react";
 import { useFocusEffect } from "@react-navigation/native";
-import { View, ScrollView } from "react-native";
+import { View, TouchableOpacity, ScrollView } from "react-native";
 import { Link } from "expo-router";
 import { useSQLiteContext } from "expo-sqlite";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
@@ -11,10 +11,21 @@ import Reanimated, {
 } from "react-native-reanimated";
 import { Button } from "~/components/ui/button";
 import { Text } from "~/components/ui/text";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "~/components/ui/alert-dialog";
 import { Token } from "~/components/Token";
 import { QrCode } from "~/lib/icons/QrCode";
 import { Trash } from "~/lib/icons/Trash";
 import { authenticator } from "~/lib/authenticator";
+import { TriangleAlert } from "~/lib/icons/TriangleAlert";
 
 type TwoFa = {
   id: number;
@@ -29,10 +40,19 @@ type RightActionProps = {
   prog: SharedValue<number>;
   drag: SharedValue<number>;
   id: number;
+  issuer: string;
   deleteTwoFa: (id: number) => void;
 };
 
-function RightAction({ prog, drag, id, deleteTwoFa }: RightActionProps) {
+function RightAction({
+  prog,
+  drag,
+  id,
+  issuer,
+  deleteTwoFa,
+}: RightActionProps) {
+  const [open, setOpen] = useState(false);
+
   const styleAnimation = useAnimatedStyle(() => {
     return {
       transform: [{ translateX: drag.value + 75 }],
@@ -41,15 +61,43 @@ function RightAction({ prog, drag, id, deleteTwoFa }: RightActionProps) {
 
   return (
     <Reanimated.View style={styleAnimation}>
-      <View className="w-[75px] bg-red-500 h-full flex justify-center items-center">
-        <Button
-          variant="destructive"
-          className="flex flex-row gap-2 w-3/5"
-          onPress={() => deleteTwoFa(id)}
-        >
-          <Trash className="text-white" />
-        </Button>
-      </View>
+      <TouchableOpacity
+        className="w-[75px] bg-red-500 h-full flex justify-center items-center"
+        onPress={() => setOpen(true)}
+      >
+        <Trash className="text-white" />
+      </TouchableOpacity>
+      <AlertDialog open={open} onOpenChange={setOpen}>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle className="text-center native:text-3xl">
+              Remove {issuer}?
+            </AlertDialogTitle>
+            <AlertDialogDescription>
+              <Text>
+                When you remove this account from PPI Authenticator, you won't
+                get codes to help you sign in securely anymore. Make sure you
+                update your settings for this account to reflect this change.
+              </Text>
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter className="flex-row justify-end">
+            <AlertDialogCancel>
+              <Text>Cancel</Text>
+            </AlertDialogCancel>
+            <AlertDialogAction asChild>
+              <Button
+                variant="destructive"
+                className="flex flex-row gap-2 bg-destructive"
+                onPress={() => deleteTwoFa(id)}
+              >
+                <Trash className="text-white" />
+                <Text>Remove Account</Text>
+              </Button>
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </Reanimated.View>
   );
 }
@@ -72,6 +120,7 @@ export default function Index() {
     await db.runAsync("DELETE FROM two_fas WHERE id = $id", {
       $id: id,
     });
+    getTwoFas();
   };
 
   useFocusEffect(
@@ -104,7 +153,7 @@ export default function Index() {
         onPress={deleteTwoFas}
       >
         <Trash className="text-white" />
-        <Text>Delete All Token</Text>
+        <Text>Remove All Account</Text>
       </Button>
       <ScrollView className="w-full">
         <GestureHandlerRootView>
@@ -120,6 +169,7 @@ export default function Index() {
                     prog={prog}
                     drag={drag}
                     id={value.id}
+                    issuer={value.issuer}
                     deleteTwoFa={deleteTwoFa}
                   />
                 )}
